optionally disable boost.chrono in duration.hpp
add BOOST_COMPUTE_NO_BOOST_CHRONO to allow user to not using boost.chrono, which is not a header-only library. boost.compute is header-only and shall allow user to remain using it header only. Including boost.chrono means that user has no choice but build boost. On the other hand, std::chrono is often header only, and which is available to most platforms today.